+++ src/test/ExSln2/c99/gen/hal_Helper.c
@@ -5,5 +5,5 @@
 
 uint8_t hal_Helper_calc_stuff(uint8_t a, uint8_t b)
 {
-    return a + b;
+    return (uint8_t)((a + b) << (b + 1));
 }